requirements.txt
# Core MCP and Groq integration
mcp[cli]==1.13.1


# FastAPI + Pydantic v2 support (needed for MCP compatibility)
pydantic-settings==2.10.1

# Uvicorn (must match MCP’s latest)

# Streamlit Dashboard

# ORM if needed later


# Environment variables
python-dotenv==1.0.1

# HTTP and SSE support for MCP tools
httpx==0.27.2
httpx-sse==0.4.1

# JSON Schema for MCP resources
jsonschema==4.22.0



# Requests for external API calls
requests==2.32.3

fastapi==0.111.0
uvicorn==0.31.1
requests==2.32.3
python-dotenv==1.0.1
pydantic>=2.11.0
streamlit==1.36.0
sqlalchemy==2.0.23
pytest==7.4.4
groq==0.9.0
